+++ lib/src/widgets/input.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
+
 import 'attachment_button.dart';
 import 'inherited_chat_theme.dart';
 import 'inherited_l10n.dart';
@@ -23,6 +24,7 @@ class Input extends StatefulWidget {
     this.isAttachmentUploading,
     this.onAttachmentPressed,
     required this.onSendPressed,
+    this.onTextChanged,
   }) : super(key: key);
 
   /// See [AttachmentButton.onPressed]
@@ -37,6 +39,9 @@ class Input extends StatefulWidget {
   /// Will be called on [SendButton] tap. Has [types.PartialText] which can
   /// be transformed to [types.TextMessage] and added to the messages list.
   final void Function(types.PartialText) onSendPressed;
+
+  /// Will be called whenever the text inside [TextField] changes
+  final void Function(types.PartialText)? onTextChanged;
 
   @override
   _InputState createState() => _InputState();
@@ -141,6 +146,9 @@ class _InputState extends State<Input> {
                     Expanded(
                       child: TextField(
                         controller: _textController,
+                        onChanged: widget.onTextChanged != null
+                            ? (text) => _onTextChanged(text)
+                            : null,
                         decoration: InputDecoration.collapsed(
                           hintStyle: InheritedChatTheme.of(context)
                               .theme
@@ -183,5 +191,10 @@ class _InputState extends State<Input> {
         ),
       ),
     );
+  }
+
+  void _onTextChanged(String? text) {
+    final partialText = types.PartialText(text: _textController.text.trim());
+    widget.onTextChanged!(partialText);
   }
 }